+++ CHANGELOG.md
@@ -7,5 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Project initialization
+- Tailwind Css added
+- Editor config added
 
 [Unreleased]: https://github.com/JuanZea/capilog-frontend